requirements.txt
aniso8601==9.0.1
async-generator==1.10
attrs==22.1.0
bleach==6.0.0
boto3==1.26.27
botocore==1.29.27
cached-property==1.5.2
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coverage==7.0.0
datapackage==1.15.2
dnspython==2.2.1
dominate==2.7.0
email-validator==1.3.0
et-xmlfile==1.1.0
exceptiongroup==1.1.0
Faker==15.3.4
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Moment==1.0.5
Flask-PageDown==0.4.0
flask-paginate==2022.1.8
Flask-RESTful==0.3.9
flask-unittest==0.1.3
Flask-WTF==1.1.1
greenlet==2.0.1
h11==0.14.0
idna==3.4
ijson==3.1.4
importlib-metadata==5.1.0
importlib-resources==5.10.1
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonlines==3.1.0
jsonpointer==2.3
jsonschema==4.17.3
linear-tsv==1.1.0
Markdown==3.4.1
MarkupSafe==2.1.1
openpyxl==3.0.10
outcome==1.2.0
packaging==22.0
peewee==3.15.4
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
pycparser==2.21
pyparsing==3.0.9
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
random-password-generator==2.2.0
requests==2.28.1
rfc3986==2.0.0
s3transfer==0.6.0
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.44
tableschema==1.20.2
tabulator==1.53.5
trio==0.22.0
trio-websocket==0.9.2
unicodecsv==0.14.1
urllib3==1.26.13
visitor==0.1.3
webencodings==0.5.1
Werkzeug==2.2.2
wsproto==1.2.0
WTForms==3.0.1
xlrd==2.0.1
zipp==3.11.0
